messaging_app/requirements.txt
amqp==5.3.1
asgiref==3.9.0
asttokens==3.0.0
billiard==4.2.1
celery==5.5.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
comm==0.2.2
debugpy==1.8.14
decorator==5.2.1
Django==5.2.4
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.16.0
drf-yasg==1.21.10
executing==2.2.0
inflection==0.5.1
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.8.1
kombu==5.5.4
matplotlib-inline==0.1.7
mysql-connector-python==9.3.0
mysqlclient==2.2.7
nest-asyncio==1.6.0
packaging==25.0
parso==0.8.4
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
pure_eval==0.2.3
Pygments==2.19.1
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==310
PyYAML==6.0.2
pyzmq==26.4.0
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
tornado==6.5.1
traitlets==5.14.3
tzdata==2025.2
uritemplate==4.2.0
vine==5.1.0
wcwidth==0.2.13
